% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ends}
\alias{ends}
\title{Dati OECD sui salari medi alla fine del periodo di osservazione.}
\format{Un data frame con 6 righe e 4 variabili:
\describe{
  \item{COUNTRY}{il paese osservato}
  \item{obsTime}{il tempo dell'osservazione}
  \item{salario_medio}{il salario medio}
  \item{pil_ora}{il pil per ora lavorata}
}}
\usage{
ends
}
\description{
I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
}
\keyword{datasets}
