% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mins}
\alias{mins}
\title{Dati OECD sull'anno in cui si è rilevato il minimo del salario medio annuo durante il periodo di osservazione 1992-2017.}
\format{Un data frame con 6 righe e 4 variabili:
\describe{
  \item{COUNTRY}{il paese osservato}
  \item{obsTime}{l'anno in cui il salario medio ha toccato il minimo in quel paese}
  \item{salario_medio}{il salario medio}
  \item{pil_ora}{il pil per ora lavorata}
}}
\usage{
mins
}
\description{
I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
Il dato è annuale ed è stato elaborato per il periodo 1992-2017.
}
\keyword{datasets}
