% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ora_lavorata}
\alias{ora_lavorata}
\title{Dati OECD sul PIL per ora lavorata durante il periodo di osservazione 1992-2017.}
\format{Un data frame con 154 righe e 3 variabili:
\describe{
  \item{LOCATION}{il paese osservato}
  \item{obsTime}{il tempo dell'osservazione}
  \item{pil_ora}{il PIL prodotto per ora di lavoro, in dollari a prezzi costanti e PPPs 2010}
}}
\usage{
ora_lavorata
}
\description{
Si tratta di una misura lorda di produttività del lavoro. I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
Il dato è annuale ed è stato elaborato per il periodo 1992-2017.
Per questa variabile non sono disponibili i dati 2017 per Giappone e USA.
}
\keyword{datasets}
