% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prod_sal}
\alias{prod_sal}
\title{Dati OECD sul PIL per ora lavorata e salario medio durante il periodo di osservazione 1992-2017.}
\format{Un data frame con 155 righe e 4 variabili:
\describe{
  \item{COUNTRY}{il paese osservato}
  \item{obsTime}{il tempo dell'osservazione}
  \item{salario_medio}{il salario medio per quel paese e quel periodo, in dollari a prezzi costanti e PPPs 2017}
  \item{pil_ora}{il PIL prodotto per ora di lavoro, in dollari a prezzi costanti e PPPs 2010}
}}
\usage{
prod_sal
}
\description{
Tabella che pone a confronto le due quantità. I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
Il dato è annuale ed è stato elaborato per il periodo 1992-2017.
}
\keyword{datasets}
