% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quarts}
\alias{quarts}
\title{Dati OECD sui quartili temporali del salario durante il periodo di osservazione 1992-2017.}
\format{Un data frame con 156 righe e 6 variabili:
\describe{
  \item{COUNTRY}{il paese osservato}
  \item{quart1}{il primo quartile della distribuzione temporale}
  \item{quart2}{il terzo quartile della distribuzione temporale}
  \item{obsTime}{il tempo dell'osservazione}
  \item{salario_medio}{l'anno in cui il salario medio ha toccato il minimo in quel paese}
  \item{pil_ora}{il pil per ora lavorata}
}}
\usage{
quarts
}
\description{
I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
Il dato è annuale ed è stato elaborato per il periodo 1992-2017.
}
\keyword{datasets}
