% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salario_medio}
\alias{salario_medio}
\title{Dati OECD sui salari medi annui durante il periodo di osservazione 1992-2017.}
\format{Un data frame con 156 righe e 3 variabili:
\describe{
  \item{COUNTRY}{il paese osservato}
  \item{obsTime}{il tempo dell'osservazione}
  \item{salario_medio}{il salario medio per quel paese e quel periodo, in dollari a prezzi costanti e PPPs 2017}
}}
\usage{
salario_medio
}
\description{
I dati sono generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
Il dato è annuale ed è stato elaborato per il periodo 1992-2017.
}
\keyword{datasets}
