% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ulc_gdp_s}
\alias{ulc_gdp_s}
\title{Dati OECD sul PIL per occupato durante il periodo di osservazione 1992-2018.}
\format{Un data frame con 666 righe e 3 variabili:
\describe{
  \item{LOCATION}{il paese osservato}
  \item{obsTime}{il tempo dell'osservazione, in questo caso trimestrale}
  \item{pil_nr_indice}{numero indice del PIL prodotto occupato, numero indice, media 2015 = 100}
}}
\usage{
ulc_gdp_s
}
\description{
Si tratta di una misura lorda di produttività del lavoro. I dati sono trimestrali e generati dal file data-raw/elabora_dati_oecd.R che elabora file scaricati dai sistemi OECD con lo script data-raw/prod_oecd_download.R
I dati si riferiscono a Italia (ITA), Francia (FRA), Giappone (JPN), Germania (DEU), Gran Bretagna (GBR) e Stati Uniti.
il dato è trimestrale ed è stato elaborato per il periodo dal 1991-Q1 al 2018-Q3.
}
\keyword{datasets}
